review comment addressed
